// lib/schemas.ts
export const GetSiteContentResponse = zod.object({
  homeTitle: zod.string(),
  homeDescription: zod.string(),
  servicesDescription: zod.string(),
  contactEmail: zod.string(),
});

export const UpdateSiteContentBody = zod.object({
  homeTitle: zod.string().optional(),
  homeDescription: zod.string().optional(),
  servicesDescription: zod.string().optional(),
  contactEmail: zod.string().optional(),
});